Pre-load proto files at import time
This prevents warning messages emitted by Home Assistant about blocking
open calls. Implementing the pre-loading required breaking import
cycles.
