+++ custom_components/ecoflow_cloud/_preload_proto.py
@@ -1,0 +1,5 @@
+from .devices.internal.proto import (
+    ecopacket_pb2,  # noqa: F401 # pyright: ignore[reportUnusedImport]
+    platform_pb2,  # noqa: F401 # pyright: ignore[reportUnusedImport]
+    powerstream_pb2,  # noqa: F401 # pyright: ignore[reportUnusedImport]
+)
